Updated calculator to use floating point numbers

+++ lesson_2/calculator.py
@@ -20,7 +20,7 @@ def prompt(message):
 
 def invalid_number(number_str):
     try:
-        int(number_str)
+        float(number_str)
     except ValueError:
         return True
 
@@ -57,13 +57,13 @@ def calculator():
 
     match operation:
         case '1':
-            output = int(number1) + int(number2)
+            output = float(number1) + int(number2)
         case '2':
-            output = int(number1) - int(number2)
+            output = float(number1) - int(number2)
         case '3':
-            output = int(number1) * int(number2)
+            output = float(number1) * int(number2)
         case '4':
-            output = int(number1) / int(number2)
+            output = float(number1) / int(number2)
